pyproject.toml
[tool.poetry]
name = "pyeodh"
version = "0.0.7"
description = "A lightweight Python client for EODH APIs."
authors = ["Dusan Figala <dusan@oxidian.com>"]
license = "BSD-2-Clause"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests-oauthlib = "^2.0.0"
pystac = "^1.10.0"
owslib = "^0.31.0"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
pyright = "^1.1.358"
ipykernel = "^6.29.4"
python-dotenv = "^1.0.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
vcrpy = "^6.0.1"
pytest-cov = "^5.0.0"
pytest-recording = "^0.13.2"
pytest-mock = "^3.14.0"

[tool.poetry.group.docs.dependencies]
sphinx = "^7.2.6"
sphinx-rtd-theme = "^2.0.0"


[tool.poetry.group.notebooks.dependencies]
shapely = "^2.0.4"
matplotlib = "^3.9.0"
pillow = "^10.3.0"
rioxarray = "^0.15.7"
fsspec = "^2024.6.1"
folium = "^0.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pyright]
include = ["pyeodh", "tests"]
# typeCheckingMode = "strict"